pyproject.toml
[project]
name = "agent-comeon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.2.13",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.1.11",
    "langchain-tavily>=0.2.17",
    "tavily-python>=0.7.23",
]
